+++ tests/core/CoreTest.php
@@ -17,7 +17,7 @@ class CoreTest extends SapphireTest {
 
 	public function testGetTempPathInProject() {
 		if(file_exists($this->tempPath)) {
-			$this->assertEquals(getTempFolder(), $this->tempPath);
+			$this->assertEquals(getTempFolder(BASE_PATH), $this->tempPath);
 		} else {
 			// A typical Windows location for where sites are stored on IIS
 			$this->assertEquals(getTempFolder('C:\\inetpub\\wwwroot\\silverstripe-test-project'), sys_get_temp_dir() . '/silverstripe-cacheC--inetpub-wwwroot-silverstripe-test-project');
